feat[st]: Factory for data of challenges
